# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Hasanaa e-commerce API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface SuccessResponse {
  success: boolean;
}

export interface LoginBody {
  username: string;
  password: string;
}

export interface LoginResponse {
  success: boolean;
  admin: string;
}

export interface AdminSession {
  authenticated: boolean;
  /** @nullable */
  admin: string | null;
}

export interface Banner {
  id: number;
  title: string;
  /** @nullable */
  subtitle: string | null;
  imageUrl: string;
  /** @nullable */
  linkUrl: string | null;
  /** @nullable */
  categoryId: number | null;
  /** @nullable */
  subcategoryId: number | null;
  position: string;
  isActive: boolean;
  sortOrder: number;
  createdAt: string;
  updatedAt: string;
}

export interface CreateBannerBody {
  title: string;
  /** @nullable */
  subtitle?: string | null;
  imageUrl: string;
  /** @nullable */
  linkUrl?: string | null;
  /** @nullable */
  categoryId?: number | null;
  /** @nullable */
  subcategoryId?: number | null;
  position?: string;
  isActive?: boolean;
  sortOrder?: number;
}

export interface UpdateBannerBody {
  /** @nullable */
  title?: string | null;
  /** @nullable */
  subtitle?: string | null;
  /** @nullable */
  imageUrl?: string | null;
  /** @nullable */
  linkUrl?: string | null;
  /** @nullable */
  categoryId?: number | null;
  /** @nullable */
  subcategoryId?: number | null;
  /** @nullable */
  position?: string | null;
  /** @nullable */
  isActive?: boolean | null;
  /** @nullable */
  sortOrder?: number | null;
}

export interface Category {
  id: number;
  name: string;
  slug: string;
  /** @nullable */
  imageUrl: string | null;
  /** @nullable */
  parentTag: string | null;
  isActive: boolean;
  sortOrder: number;
  createdAt: string;
  updatedAt: string;
}

export interface Subcategory {
  id: number;
  name: string;
  slug: string;
  categoryId: number;
  /** @nullable */
  imageUrl: string | null;
  isActive: boolean;
  sortOrder: number;
  createdAt: string;
  updatedAt: string;
}

export interface Product {
  id: number;
  name: string;
  /** @nullable */
  description: string | null;
  price: string;
  /** @nullable */
  originalPrice: string | null;
  /** @nullable */
  imageUrl: string | null;
  images: string[];
  /** @nullable */
  categoryId: number | null;
  /** @nullable */
  categoryName: string | null;
  /** @nullable */
  subcategoryId: number | null;
  /** @nullable */
  subcategoryName: string | null;
  inStock: boolean;
  isFeatured: boolean;
  sizes: string[];
  colors: string[];
  createdAt: string;
  updatedAt: string;
}

export interface CategoryWithProducts {
  id: number;
  name: string;
  slug: string;
  /** @nullable */
  imageUrl: string | null;
  /** @nullable */
  parentTag: string | null;
  isActive: boolean;
  sortOrder: number;
  createdAt: string;
  updatedAt: string;
  subcategories: Subcategory[];
  products: Product[];
}

export interface CreateCategoryBody {
  name: string;
  slug: string;
  /** @nullable */
  imageUrl?: string | null;
  /** @nullable */
  parentTag?: string | null;
  isActive?: boolean;
  sortOrder?: number;
}

export interface UpdateCategoryBody {
  /** @nullable */
  name?: string | null;
  /** @nullable */
  slug?: string | null;
  /** @nullable */
  imageUrl?: string | null;
  /** @nullable */
  parentTag?: string | null;
  /** @nullable */
  isActive?: boolean | null;
  /** @nullable */
  sortOrder?: number | null;
}

export interface SubcategoryWithProducts {
  id: number;
  name: string;
  slug: string;
  categoryId: number;
  /** @nullable */
  imageUrl: string | null;
  isActive: boolean;
  sortOrder: number;
  createdAt: string;
  updatedAt: string;
  products: Product[];
}

export interface CreateSubcategoryBody {
  name: string;
  slug: string;
  categoryId: number;
  /** @nullable */
  imageUrl?: string | null;
  isActive?: boolean;
  sortOrder?: number;
}

export interface UpdateSubcategoryBody {
  /** @nullable */
  name?: string | null;
  /** @nullable */
  slug?: string | null;
  /** @nullable */
  categoryId?: number | null;
  /** @nullable */
  imageUrl?: string | null;
  /** @nullable */
  isActive?: boolean | null;
  /** @nullable */
  sortOrder?: number | null;
}

export interface PromotionItem {
  imageUrl: string;
  label: string;
  /** @nullable */
  categoryId?: number | null;
  /** @nullable */
  subcategoryId?: number | null;
}

export interface Promotion {
  id: number;
  title: string;
  /** 2 for 2-column grid, 4 for 4-column grid */
  gridType: number;
  /** top or bottom */
  position: string;
  isActive: boolean;
  sortOrder: number;
  items: PromotionItem[];
  createdAt: string;
  updatedAt: string;
}

export interface CreatePromotionBody {
  title: string;
  gridType: number;
  position: string;
  isActive?: boolean;
  sortOrder?: number;
  items?: PromotionItem[];
}

export interface UpdatePromotionBody {
  /** @nullable */
  title?: string | null;
  /** @nullable */
  gridType?: number | null;
  /** @nullable */
  position?: string | null;
  /** @nullable */
  isActive?: boolean | null;
  /** @nullable */
  sortOrder?: number | null;
  items?: PromotionItem[];
}

export interface CreateProductBody {
  name: string;
  /** @nullable */
  description?: string | null;
  price: string;
  /** @nullable */
  originalPrice?: string | null;
  /** @nullable */
  imageUrl?: string | null;
  images?: string[];
  /** @nullable */
  categoryId?: number | null;
  /** @nullable */
  subcategoryId?: number | null;
  inStock?: boolean;
  isFeatured?: boolean;
  sizes?: string[];
  colors?: string[];
}

export interface UpdateProductBody {
  /** @nullable */
  name?: string | null;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  price?: string | null;
  /** @nullable */
  originalPrice?: string | null;
  /** @nullable */
  imageUrl?: string | null;
  images?: string[];
  /** @nullable */
  categoryId?: number | null;
  /** @nullable */
  subcategoryId?: number | null;
  /** @nullable */
  inStock?: boolean | null;
  /** @nullable */
  isFeatured?: boolean | null;
  sizes?: string[];
  colors?: string[];
}

export interface OrderItem {
  productId: number;
  productName: string;
  price: string;
  quantity: number;
  /** @nullable */
  size: string | null;
  /** @nullable */
  color: string | null;
}

export interface Order {
  id: number;
  customerName: string;
  customerPhone: string;
  customerAddress: string;
  /** @nullable */
  customerCity: string | null;
  /** @nullable */
  notes: string | null;
  status: string;
  totalAmount: string;
  items: OrderItem[];
  createdAt: string;
  updatedAt: string;
}

export interface CreateOrderItemBody {
  productId: number;
  quantity: number;
  /** @nullable */
  size?: string | null;
  /** @nullable */
  color?: string | null;
}

export interface CreateOrderBody {
  customerName: string;
  customerPhone: string;
  customerAddress: string;
  /** @nullable */
  customerCity?: string | null;
  /** @nullable */
  notes?: string | null;
  items: CreateOrderItemBody[];
}

export interface UpdateOrderStatusBody {
  status: string;
}

export interface OrderStats {
  totalOrders: number;
  pendingOrders: number;
  confirmedOrders: number;
  deliveredOrders: number;
  cancelledOrders: number;
  totalRevenue: string;
  recentOrders: Order[];
}

export interface UploadImageBody {
  /** Base64 encoded image data with data URI prefix */
  data: string;
  filename: string;
}

export interface UploadImageResponse {
  url: string;
}

export type ListSubcategoriesParams = {
  /**
   * @nullable
   */
  categoryId?: number | null;
};

export type ListProductsParams = {
  /**
   * @nullable
   */
  categoryId?: number | null;
  /**
   * @nullable
   */
  subcategoryId?: number | null;
  /**
   * @nullable
   */
  featured?: string | null;
};