# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Hasanaa e-commerce API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /banners:
    get:
      operationId: listBanners
      tags: [banners]
      summary: List all banners
      responses:
        "200":
          description: List of banners
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Banner"
    post:
      operationId: createBanner
      tags: [banners]
      summary: Create a banner
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateBannerBody"
      responses:
        "201":
          description: Banner created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Banner"
  /banners/{id}:
    get:
      operationId: getBanner
      tags: [banners]
      summary: Get a banner
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Banner
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Banner"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: updateBanner
      tags: [banners]
      summary: Update a banner
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateBannerBody"
      responses:
        "200":
          description: Banner updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Banner"
    delete:
      operationId: deleteBanner
      tags: [banners]
      summary: Delete a banner
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Banner deleted
  /categories:
    get:
      operationId: listCategories
      tags: [categories]
      summary: List all categories
      responses:
        "200":
          description: List of categories
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Category"
    post:
      operationId: createCategory
      tags: [categories]
      summary: Create a category
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCategoryBody"
      responses:
        "201":
          description: Category created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Category"
  /categories/{id}:
    get:
      operationId: getCategory
      tags: [categories]
      summary: Get a category with its products and subcategories
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Category
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CategoryWithProducts"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: updateCategory
      tags: [categories]
      summary: Update a category
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateCategoryBody"
      responses:
        "200":
          description: Category updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Category"
    delete:
      operationId: deleteCategory
      tags: [categories]
      summary: Delete a category
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Category deleted
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
...
    Banner:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        subtitle:
          type: ["string", "null"]
        imageUrl:
          type: string
        linkUrl:
          type: ["string", "null"]
        categoryId:
          type: ["integer", "null"]
        subcategoryId:
          type: ["integer", "null"]
        position:
          type: string
        isActive:
          type: boolean
        sortOrder:
          type: integer
        createdAt:
          type: string
        updatedAt:
          type: string
      required:
        - id
        - title
        - subtitle
        - imageUrl
        - linkUrl
        - categoryId
        - subcategoryId
        - position
        - isActive
        - sortOrder
        - createdAt
        - updatedAt
    CreateBannerBody:
      type: object
      properties:
        title:
          type: string
        subtitle:
          type: ["string", "null"]
        imageUrl:
          type: string
        linkUrl:
          type: ["string", "null"]
        categoryId:
          type: ["integer", "null"]
        subcategoryId:
          type: ["integer", "null"]
        position:
          type: string
        isActive:
          type: boolean
        sortOrder:
          type: integer
      required:
        - title
        - imageUrl
    UpdateBannerBody:
      type: object
      properties:
        title:
          type: ["string", "null"]
        subtitle:
          type: ["string", "null"]
        imageUrl:
          type: ["string", "null"]
        linkUrl:
          type: ["string", "null"]
        categoryId:
          type: ["integer", "null"]
        subcategoryId:
          type: ["integer", "null"]
        position:
          type: ["string", "null"]
        isActive:
          type: ["boolean", "null"]
        sortOrder:
          type: ["integer", "null"]
    Category:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        slug:
          type: string
        imageUrl:
          type: ["string", "null"]
        parentTag:
          type: ["string", "null"]
        isActive:
          type: boolean
        sortOrder:
          type: integer
        createdAt:
          type: string
        updatedAt:
          type: string
      required:
        - id
        - name
        - slug
        - imageUrl
        - parentTag
        - isActive
        - sortOrder
        - createdAt
        - updatedAt
    CategoryWithProducts:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        slug:
          type: string
        imageUrl:
          type: ["string", "null"]
        parentTag:
          type: ["string", "null"]
        isActive:
          type: boolean
        sortOrder:
          type: integer
        createdAt:
          type: string
        updatedAt:
          type: string
        subcategories:
          type: array
          items:
            $ref: "#/components/schemas/Subcategory"
        products:
          type: array
          items:
            $ref: "#/components/schemas/Product"
      required:
        - id
        - name
        - slug
        - imageUrl
        - parentTag
        - isActive
        - sortOrder
        - createdAt
        - updatedAt
        - subcategories
        - products
    CreateCategoryBody:
      type: object
      properties:
        name:
          type: string
        slug:
          type: string
        imageUrl:
          type: ["string", "null"]
        parentTag:
          type: ["string", "null"]
        isActive:
          type: boolean
        sortOrder:
          type: integer
      required:
        - name
        - slug
    UpdateCategoryBody:
      type: object
      properties:
        name:
          type: ["string", "null"]
        slug:
          type: ["string", "null"]
        imageUrl:
          type: ["string", "null"]
        parentTag:
          type: ["string", "null"]
        isActive:
          type: ["boolean", "null"]
        sortOrder:
          type: ["integer", "null"]
    Subcategory:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        slug:
          type: string
        categoryId:
          type: integer
        imageUrl:
          type: ["string", "null"]
        isActive:
          type: boolean
        sortOrder:
          type: integer
        createdAt:
          type: string
        updatedAt:
          type: string
      required:
        - id
        - name
        - slug
        - categoryId
        - imageUrl
        - isActive
        - sortOrder
        - createdAt
        - updatedAt
...
    Product:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: ["string", "null"]
        price:
          type: string
        originalPrice:
          type: ["string", "null"]
        imageUrl:
          type: ["string", "null"]
        images:
          type: array
          items:
            type: string
        categoryId:
          type: ["integer", "null"]
        categoryName:
          type: ["string", "null"]
        subcategoryId:
          type: ["integer", "null"]
        subcategoryName:
          type: ["string", "null"]
        inStock:
          type: boolean
        isFeatured:
          type: boolean
        sizes:
          type: array
          items:
            type: string
        colors:
          type: array
          items:
            type: string
        createdAt:
          type: string
        updatedAt:
          type: string
      required:
        - id
        - name
        - description
        - price
        - originalPrice
        - imageUrl
        - images
        - categoryId
        - categoryName
        - subcategoryId
        - subcategoryName
        - inStock
        - isFeatured
        - sizes
        - colors
        - createdAt
        - updatedAt
...